# Execute stream callbacks directly when executor is the current event …