…loop.

TBH, haven't been able to reason through whether this is a correct change, but tests seem fine, and the benchmarks show that we need to do something along these lines to optimize `StreamMessage` performance for the standard case where everything happens on a server event loop thread. Curious what you think.

Benchmarks (due to synchronization overhead, only high values of `num` are relevant). Probably about 1.5-2x better stream performance in practice.

After:
```
Benchmark                                                    (num)   Mode  Cnt       Score       Error  Units
StreamMessageBenchmark.defaultStreamMessage_notJmhEventLoop      3  thrpt   20  238282.293 ± 11059.030  ops/s
StreamMessageBenchmark.defaultStreamMessage_notJmhEventLoop      5  thrpt   20  223524.752 ±  5595.936  ops/s
StreamMessageBenchmark.defaultStreamMessage_notJmhEventLoop     20  thrpt   20  170606.288 ±  4851.481  ops/s
StreamMessageBenchmark.defaultStreamMessage_notJmhEventLoop    100  thrpt   20   78263.884 ±  1800.286  ops/s
StreamMessageBenchmark.defaultStreamMessage_notJmhEventLoop   1000  thrpt   20   10792.808 ±   216.650  ops/s
```

Before:
```
Benchmark                                                    (num)   Mode  Cnt       Score      Error  Units
StreamMessageBenchmark.defaultStreamMessage_notJmhEventLoop      3  thrpt   20  222284.597 ± 6456.982  ops/s
StreamMessageBenchmark.defaultStreamMessage_notJmhEventLoop      5  thrpt   20  215316.254 ± 6570.347  ops/s
StreamMessageBenchmark.defaultStreamMessage_notJmhEventLoop     20  thrpt   20  173076.370 ± 5739.022  ops/s
StreamMessageBenchmark.defaultStreamMessage_notJmhEventLoop    100  thrpt   20   60886.142 ± 1018.291  ops/s
StreamMessageBenchmark.defaultStreamMessage_notJmhEventLoop   1000  thrpt   20    6617.666 ±  118.178  ops/s
```