@anuraaga 2017-11-17
I have split off a PR to first solve the reentrancy problem in the non-executor case which seems required by the spec.